Move convert_ambiguous_timestamp_to_ns
Put it next to the queue time tools and out of the ambiguously named 'util' module.
