+++ src/scout_apm/core/queue_time.py
@@ -1,8 +1,9 @@
 # coding=utf-8
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+import time
+
 from scout_apm.compat import datetime_to_timestamp
-from scout_apm.core.util import convert_ambiguous_timestamp_to_ns
 
 
 def track_request_queue_time(header_value, tracked_request):
@@ -34,3 +35,30 @@ def track_request_queue_time(header_value, tracked_request):
 
     queue_time_ns = int(tr_start_timestamp_ns - start_timestamp_ns)
     tracked_request.tag("scout.queue_time_ns", queue_time_ns)
+
+
+# Cutoff epoch is used for determining ambiguous timestamp boundaries, and is
+# just over 10 years ago at time of writing
+CUTOFF_EPOCH_S = time.mktime((2009, 6, 1, 0, 0, 0, 0, 0, 0))
+CUTOFF_EPOCH_MS = CUTOFF_EPOCH_S * 1000.0
+CUTOFF_EPOCH_US = CUTOFF_EPOCH_S * 1000000.0
+CUTOFF_EPOCH_NS = CUTOFF_EPOCH_S * 1000000000.0
+
+
+def convert_ambiguous_timestamp_to_ns(timestamp):
+    """
+    Convert an ambiguous float timestamp that could be in nanoseconds,
+    microseconds, milliseconds, or seconds to nanoseconds. Return 0.0 for
+    values in the more than 10 years ago.
+    """
+    if timestamp > CUTOFF_EPOCH_NS:
+        converted_timestamp = timestamp
+    elif timestamp > CUTOFF_EPOCH_US:
+        converted_timestamp = timestamp * 1000.0
+    elif timestamp > CUTOFF_EPOCH_MS:
+        converted_timestamp = timestamp * 1000000.0
+    elif timestamp > CUTOFF_EPOCH_S:
+        converted_timestamp = timestamp * 1000000000.0
+    else:
+        return 0.0
+    return converted_timestamp
